renamed all occurence of appNotifier to notifier
