+++ controllers/comment.js
@@ -1,6 +1,6 @@
 import Sequelize from 'sequelize';
 import { format } from 'date-fns';
-import { AppNotifier } from '../services';
+import { Notifier } from '../services';
 import models from '../models';
 
 const { iLike } = Sequelize.Op;
@@ -20,7 +20,7 @@ export default {
 
     const comment = Object.values(newComment.comment)[0];
     const { id, updatedAt } = newComment;
-    await AppNotifier.notifyInApp(article.id, decoded.id);
+    await Notifier.notifyInApp(article.id, decoded.id);
 
     return res.json({ comment, id, updatedAt });
   },
