+++ services/index.js
@@ -1,2 +1,2 @@
 export { default } from './emailService';
-export { default as AppNotifier } from './appNotifier';
+export { default as Notifier } from './notifier';
